feat: adds Response Entity to controllers

+++ src/main/java/dev/java10x/employee_management/Functionarys/FunctionaryController.java
@@ -1,5 +1,7 @@
 package dev.java10x.employee_management.Functionarys;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,28 +18,44 @@ public class FunctionaryController {
     }
 
     @GetMapping("/list")
-    public List<FunctionaryDTO> list() {
-        return functionaryService.list();
+    public ResponseEntity<List<FunctionaryDTO>> list() {
+        List<FunctionaryDTO> functionarys = functionaryService.list();
+        return ResponseEntity.ok(functionarys);
     }
 
     @GetMapping("/list/{id}")
-    public FunctionaryDTO listById(@PathVariable long id) {
-        return functionaryService.listById(id);
+    public ResponseEntity<?> listById(@PathVariable long id) {
+        FunctionaryDTO functionary = functionaryService.listById(id);
+        if (functionary != null) {
+            return ResponseEntity.ok(functionary);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Employee not found!");
+        }
     }
 
     @PostMapping("/register")
-    public FunctionaryDTO register(@RequestBody FunctionaryDTO functionaryDTO) {
-        return functionaryService.register(functionaryDTO);
+    public ResponseEntity<String> register(@RequestBody FunctionaryDTO functionaryDTO) {
+        FunctionaryDTO functionary = functionaryService.register(functionaryDTO);
+        return ResponseEntity.status(HttpStatus.CREATED).body("Registered employee " + functionary.getName() + " ID: " + functionary.getId());
     }
 
     @DeleteMapping("/delete/{id}")
-    public String delete(@PathVariable long id) {
-        functionaryService.delete(id);
-        return "Functionary deleted!";
+    public ResponseEntity<String> delete(@PathVariable long id) {
+        if (functionaryService.listById(id) != null) {
+            functionaryService.delete(id);
+            return ResponseEntity.ok("Deleted employee: " + id);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Employee not found!");
+        }
     }
 
     @PatchMapping("/update/{id}")
-    public FunctionaryDTO update(@PathVariable Long id, @RequestBody Map<String, Object> fields) {
-        return functionaryService.updated(id, fields);
+    public ResponseEntity<String> update(@PathVariable Long id, @RequestBody Map<String, Object> fields) {
+        if (functionaryService.listById(id) != null) {
+            functionaryService.updated(id, fields);
+            return ResponseEntity.ok("Updated id: " + id + " Fields: " + fields);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Employee not found!");
+        }
     }
 }
